deepface-sample-mongo-flask-rest-improved/requirements.txt
Flask
python-dotenv
flask-smorest
pymongo
Flask-PyMongo
deepface
Flask-WTF